Fix all the lint errors

+++ lib/onnx/nodes/concat.ts
@@ -1,15 +1,22 @@
-import Tensor from "../../types";
-import { OnnxNode } from "../node";
-import { Attributes, Constants } from "../types";
+import Tensor from '../../types';
+import {OnnxNode} from '../node';
+import {Attributes, Constants} from '../types';
 
 export class ConcatNode extends OnnxNode {
-  private axis: number;
+  private axis?: number;
 
-  constructor(attributes: Attributes, inputs: string[], outputs: string[], constants: Constants, onnxVersion: number) {
+  constructor(
+    attributes: Attributes,
+    inputs: string[],
+    outputs: string[],
+    constants: Constants,
+    onnxVersion: number
+  ) {
     super(attributes, inputs, outputs, constants, onnxVersion);
 
     if (onnxVersion < 13) {
-      this.axis = this.getAttributeInt("axis");
+      //@ts-ignore
+      this.axis = this.getAttributeInt('axis');
     }
   }
 
@@ -19,16 +26,21 @@ export class ConcatNode extends OnnxNode {
       //console.warn(`Concat with more than 2 tensors is currently slow. Doing concat with ${inputs.length} tensors`);
     }
 
-    let result = inputs[0];
-    for (let i = 1; i < inputs.length; i++) {
-      let newRes = result.concat(inputs[i], this.axis);
-      if (i > 1) {
-        result.delete();
+    if (this.onnxVersion < 13 && this.axis !== undefined) {
+      let result = inputs[0];
+      for (let i = 1; i < inputs.length; i++) {
+        const newRes = result.concat(inputs[i], this.axis);
+        if (i > 1) {
+          result.delete();
+        }
+        result = newRes;
       }
-      result = newRes;
-    }
 
-    return [result];
+      return [result];
+    }
+    throw new Error(
+      `Concat not implemented for onnx version ${this.onnxVersion}`
+    );
   }
 
   getType() {
