+++ lib/onnx/nodes/constant.ts
@@ -1,37 +1,53 @@
-import Tensor, { Precision } from "../../types";
-import { toCPU, toGPU, toWASM } from "../../util/convert";
-import { OnnxNode } from "../node";
-import { Attributes, Constants } from "../types";
-import { createTensor } from "../util";
+import Tensor, {Precision} from '../../types';
+import {toCPU, toGPU, toWASM} from '../../util/convert';
+import {OnnxNode} from '../node';
+import {Attributes, Constants} from '../types';
+import {createTensor} from '../util';
 
 export class ConstantNode extends OnnxNode {
-  public tensor: Tensor;
+  public tensor?: Tensor;
 
-  constructor(attributes: Attributes, inputs: string[], outputs: string[], constants: Constants, onnxVersion: number) {
+  constructor(
+    attributes: Attributes,
+    inputs: string[],
+    outputs: string[],
+    constants: Constants,
+    onnxVersion: number
+  ) {
     super(attributes, inputs, outputs, constants, onnxVersion);
 
     if (onnxVersion < 11) {
-      const tensor = this.getAttributeTensor("value");
-      this.tensor = createTensor(tensor);
+      const tensor = this.getAttributeTensor('value');
+      if (tensor !== undefined && tensor !== null) {
+        this.tensor = createTensor(tensor);
+      }
     }
   }
 
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   async forward(inputs: Tensor[]): Promise<Tensor[]> {
-    if (this.onnxVersion < 11) {
+    if (this.onnxVersion < 11 && this.tensor !== undefined) {
       return [this.tensor];
     }
     throw new Error('Constant with onnx version >= 11 not yet implemented');
   }
 
   async toCPU() {
-    this.tensor = await toCPU(this.tensor);
+    if (this.tensor !== undefined) {
+      this.tensor = await toCPU(this.tensor);
+    }
   }
 
   async toWASM() {
-    this.tensor = await toWASM(this.tensor);
+    if (this.tensor !== undefined) {
+      this.tensor = await toWASM(this.tensor);
+    }
   }
+
   async toGPU(precision: Precision) {
-    this.tensor = await toGPU(this.tensor, precision);
+    if (this.tensor !== undefined) {
+      this.tensor = await toGPU(this.tensor, precision);
+    }
   }
 
   getType() {
@@ -39,6 +55,8 @@ export class ConstantNode extends OnnxNode {
   }
 
   delete(): void {
-    this.tensor.delete();
+    if (this.tensor !== undefined) {
+      this.tensor.delete();
+    }
   }
 }
