+++ lib/onnx/nodes/constantOfShape.ts
@@ -1,37 +1,50 @@
-import { CPUTensor } from "../../tensor/cpu/tensor";
-import Tensor from "../../types";
-import { getSize } from "../../util/shape";
-import { OnnxNode } from "../node";
-import { Attributes, Constants } from "../types";
-import { createTensor } from "../util";
+import {CPUTensor} from '../../tensor/cpu/tensor';
+import Tensor from '../../types';
+import {getSize} from '../../util/shape';
+import {OnnxNode} from '../node';
+import {Attributes, Constants} from '../types';
+import {createTensor} from '../util';
 
 export class ConstantOfShapeNode extends OnnxNode {
-  private tensor: CPUTensor;
+  private tensor?: CPUTensor;
 
-  constructor(attributes: Attributes, inputs: string[], outputs: string[], constants: Constants, onnxVersion: number) {
+  constructor(
+    attributes: Attributes,
+    inputs: string[],
+    outputs: string[],
+    constants: Constants,
+    onnxVersion: number
+  ) {
     super(attributes, inputs, outputs, constants, onnxVersion);
 
     if (onnxVersion < 11) {
-      const tensor = this.getAttributeTensor("value");
-      this.tensor = createTensor(tensor);
+      const tensor = this.getAttributeTensor('value');
+      if (tensor !== null && tensor !== undefined) {
+        this.tensor = createTensor(tensor);
+      }
     }
   }
 
   async forward(inputs: Tensor[]): Promise<Tensor[]> {
     const _shape = inputs[0];
 
-    if (!(_shape instanceof CPUTensor)) {
-      throw new Error("ConstantOfShape needs cpu tensor as shape tensor");
-    }
-    const shape = new Array(_shape.size);
-    for (let i = 0; i < _shape.size; i++) {
-      shape[i] = _shape.get(i);
-    }
+    if (this.onnxVersion < 11 && this.tensor !== undefined) {
+      if (!(_shape instanceof CPUTensor)) {
+        throw new Error('ConstantOfShape needs cpu tensor as shape tensor');
+      }
+      const shape = new Array(_shape.size);
+      for (let i = 0; i < _shape.size; i++) {
+        shape[i] = _shape.get(i);
+      }
 
-    const size = getSize(shape);
-    const values = new Float32Array(size).fill(this.tensor.get(0));
+      const size = getSize(shape);
+      const values = new Float32Array(size).fill(this.tensor.get(0));
 
-    return [new CPUTensor(shape, values, this.tensor.type)];
+      return [new CPUTensor(shape, values, this.tensor.type)];
+    }
+    throw new Error(
+      `ConstantOfShape not implemented for onnx version ${this.onnxVersion}`
+    );
   }
 
   getType() {
@@ -39,6 +52,8 @@ export class ConstantOfShapeNode extends OnnxNode {
   }
 
   delete(): void {
-    this.tensor.delete();
+    if (this.tensor !== undefined) {
+      this.tensor.delete();
+    }
   }
 }
